pushing interfaces

+++ CabApplication/src/main/java/cabapplication/repository/ITripRepository.java
@@ -9,8 +9,8 @@ import cabapplication.entity.TripBooking;
 
 public interface ITripRepository extends JpaRepository<TripBooking, Integer> {
 
-	public List<TripBooking> getByCustomerId(int customerId);
+	public List<TripBooking> getByCustomerId(double customerId);
 
-	public TripBooking findByCustomerId(int customerId);
+	public TripBooking findByCustomerId(double customerId);
 
 }
